+++ doc/RELEASE-NOTES.md
@@ -12,7 +12,9 @@
 * Remove 'persistentFiles' and 'diskMB' from `protocol.data.Application`
   This is incompatible change to the client-server protocol and
   client library interface.
-* Add '--tag' option to 'list' and add'killall' subcommand .
+* Add '--tag' and '--state' option to 'list' and add 'killall' subcommand .
+  This involves API change of client library and protocol change,
+  especially `ListJobRequest` changed to 'POST' request.
 * Add a set of String 'tags' to `Job` class, to selectively list jobs by a tag.
 * Change `protocol.data.Job` class, having `ResourceQuantity` as member `resources`.
   This introduces incompatible change between 0.1 and 0.2, too.